Swagger definition and endpoint to create exercises

+++ internal/database/models/question.go
@@ -1,8 +1,8 @@
 package models
 
 type Question struct {
-	QuestionID   string `json:"question_id"`
-	ExerciseID   string `json:"exercise_id"`
-	QuestionText string `json:"question_text"`
-	Points       int    `json:"points"`
+	QuestionID   string  `json:"question_id"`
+	ExerciseID   string  `json:"exercise_id"`
+	QuestionText string  `json:"question_text"`
+	Points       float32 `json:"points"`
 }
